requirements.txt
Django==3.2.5
django-simple-captcha==0.5.14
flake8==3.9.2
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
